script/etl/requirements.txt
polars
connectorx
pyarrow
black
python-dotenv